requirements.txt
asgiref==3.6.0
astroid==2.14.2
certifi==2022.12.7
charset-normalizer==3.0.1
dill==0.3.6
dj-database-url==1.2.0
Django==4.1.6
djangorestframework==3.14.0
Faker==16.7.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.0.0
psycopg2==2.9.4
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.14
whitenoise==6.3.0
wrapt==1.14.1
